refactor: move Instance testing to test_instance.py
